requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.2
boto3==1.15.14
botocore==1.18.14
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.7.0
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
isort==5.5.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pipenv==2020.11.15
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
spotipy==2.16.0
sqlparse==0.3.1
toml==0.10.1
twilio==6.45.4
urllib3==1.25.10
virtualenv==20.4.3
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1